requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.6

# OpenAI (GPT-4o Vision + Whisper)
openai>=1.0.0

# Web-Dashboard
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0

# Datenbank
sqlalchemy>=2.0.0

# Word-Export
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# PDF-Diagramme in Word (optional)
# matplotlib>=3.8.0
